move jail backups to distinct class
