+++ iocage/lib/Jail.py
@@ -871,42 +871,6 @@ class JailGenerator(JailResource):
         self.config['basejail_type'] = template.config['basejail_type']
         self._create_from_resource(template)
 
-    def export(self, destination: str) -> None:
-        """
-        Export the jail.
-
-        Jail exports contain the jails configuration and differing files
-        between the jails root dataset and its release. Other datasets in the
-        jails dataset are snapshotted and entirely attached to the export.
-
-        Args:
-
-            destination (str):
-
-                The jail is exported to this location.
-        """
-        temp_dir = tempfile.TemporaryDirectory()
-        temp_config = iocage.lib.Config.Type.JSON.ConfigJSON(
-            file=f"{temp_dir.name}/config.json",
-            logger=self.logger
-        )
-        temp_config.data = self.config.data
-        temp_config.write(temp_config.data)
-        self.logger.verbose(f"Duplicating jail config to {temp_config.file}")
-
-        temp_root_dir = f"{temp_dir.name}/root"
-        os.mkdir(temp_root_dir)
-
-        self.logger.verbose(f"Writing jail delta to {temp_root_dir}")
-        iocage.lib.helpers.exec([
-            "rsync",
-            "-rv",
-            "--checksum",
-            f"--compare-dest={self.release.root_dataset.mountpoint}/",
-            f"{self.root_dataset.mountpoint}/",
-            temp_root_dir
-        ])
-
     def promote(self) -> None:
         """Promote all datasets of the jail."""
         self.zfs.promote_dataset(self.dataset)
